+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `WP 404 Caching` will be documented in this file.
 
 - Fix issue where a empty 404 page would be returned when the cache is not yet
   set. The 404 page will now function normally until the cache is set.
+- Add default 'Cache-Control' header to the 404 response.
 - Bumped minimum PHP version to 8.2.
 
 ## 1.1.0